Centralized a number of Android and iOS document controls, moving them into /infrastructure
