+++ super_editor/lib/src/infrastructure/document_gestures.dart
