+++ super_editor/lib/src/infrastructure/platforms/android/android_document_controls.dart
@@ -1,0 +1,153 @@
+import 'dart:async';
+
+import 'package:flutter/rendering.dart';
+import 'package:super_editor/src/infrastructure/platforms/mobile_documents.dart';
+
+/// Controls the display of drag handles, a magnifier, and a
+/// floating toolbar, assuming Android-style behavior for the
+/// handles.
+class AndroidDocumentGestureEditingController extends MagnifierAndToolbarController {
+  AndroidDocumentGestureEditingController({
+    required LayerLink documentLayoutLink,
+    required LayerLink magnifierFocalPointLink,
+  })  : _documentLayoutLink = documentLayoutLink,
+        super(magnifierFocalPointLink: magnifierFocalPointLink);
+
+  @override
+  void dispose() {
+    _collapsedHandleAutoHideTimer?.cancel();
+    super.dispose();
+  }
+
+  /// Layer link that's aligned to the top-left corner of the document layout.
+  ///
+  /// Some of the offsets reported by this controller are based on the
+  /// document layout coordinate space. Therefore, to honor those offsets on
+  /// the screen, this `LayerLink` should be used to align the controls with
+  /// the document layout before applying the offset that sits within the
+  /// document layout.
+  LayerLink get documentLayoutLink => _documentLayoutLink;
+  final LayerLink _documentLayoutLink;
+
+  /// Whether or not a caret should be displayed.
+  bool get hasCaret => caretTop != null;
+
+  /// The offset of the top of the caret, or `null` if no caret should
+  /// be displayed.
+  ///
+  /// When the caret is drawn, the caret will have a thickness. That width
+  /// should be placed either on the left or right of this offset, based on
+  /// whether the [caretAffinity] is upstream or downstream, respectively.
+  Offset? get caretTop => _caretTop;
+  Offset? _caretTop;
+
+  /// The height of the caret, or `null` if no caret should be displayed.
+  double? get caretHeight => _caretHeight;
+  double? _caretHeight;
+
+  /// Updates the caret's size and position.
+  ///
+  /// The [top] offset is in the document layout's coordinate space.
+  void updateCaret({
+    Offset? top,
+    double? height,
+  }) {
+    bool changed = false;
+    if (top != null) {
+      _caretTop = top;
+      changed = true;
+    }
+    if (height != null) {
+      _caretHeight = height;
+      changed = true;
+    }
+
+    if (changed) {
+      notifyListeners();
+    }
+  }
+
+  /// Removes the caret from the display.
+  void removeCaret() {
+    if (!hasCaret) {
+      return;
+    }
+
+    _caretTop = null;
+    _caretHeight = null;
+    notifyListeners();
+  }
+
+  /// Whether a collapsed handle should be displayed.
+  bool get shouldDisplayCollapsedHandle => _collapsedHandleOffset != null;
+
+  /// The offset of the collapsed handle focal point, within the coordinate space
+  /// of the document layout, or `null` if no collapsed handle should be displayed.
+  Offset? get collapsedHandleOffset => _collapsedHandleOffset;
+  Offset? _collapsedHandleOffset;
+  set collapsedHandleOffset(Offset? offset) {
+    if (offset != _collapsedHandleOffset) {
+      _collapsedHandleOffset = offset;
+      notifyListeners();
+    }
+  }
+
+  /// Whether the expanded handles (base + extent) should be displayed.
+  bool get shouldDisplayExpandedHandles => _upstreamHandleOffset != null && _downstreamHandleOffset != null;
+
+  /// The offset of the upstream handle focal point, within the coordinate space
+  /// of the document layout, or `null` if no upstream handle should be displayed.
+  Offset? get upstreamHandleOffset => _upstreamHandleOffset;
+  Offset? _upstreamHandleOffset;
+  set upstreamHandleOffset(Offset? offset) {
+    if (offset != _upstreamHandleOffset) {
+      _upstreamHandleOffset = offset;
+      notifyListeners();
+    }
+  }
+
+  /// The offset of the downstream handle focal point, within the coordinate space
+  /// of the document layout, or `null` if no downstream handle should be displayed.
+  Offset? get downstreamHandleOffset => _downstreamHandleOffset;
+  Offset? _downstreamHandleOffset;
+  set downstreamHandleOffset(Offset? offset) {
+    if (offset != _downstreamHandleOffset) {
+      _downstreamHandleOffset = offset;
+      notifyListeners();
+    }
+  }
+
+  final Duration _collapsedHandleAutoHideDuration = const Duration(seconds: 4);
+  Timer? _collapsedHandleAutoHideTimer;
+
+  /// Whether the collapsed handle should be faded out for the purpose of
+  /// auto-hiding while the user is inactive.
+  bool get isCollapsedHandleAutoHidden => _isCollapsedHandleAutoHidden;
+  bool _isCollapsedHandleAutoHidden = false;
+
+  /// Starts a countdown that, if reached, fades out the collapsed drag handle.
+  void startCollapsedHandleAutoHideCountdown() {
+    _collapsedHandleAutoHideTimer?.cancel();
+    _collapsedHandleAutoHideTimer = Timer(_collapsedHandleAutoHideDuration, _hideCollapsedHandle);
+  }
+
+  /// Cancels a countdown that started with [startCollapsedHandleAutoHideCountdown].
+  void cancelCollapsedHandleAutoHideCountdown() {
+    _collapsedHandleAutoHideTimer?.cancel();
+  }
+
+  void _hideCollapsedHandle() {
+    if (!_isCollapsedHandleAutoHidden) {
+      _isCollapsedHandleAutoHidden = true;
+      notifyListeners();
+    }
+  }
+
+  /// Brings back a faded-out collapsed drag handle.
+  void unHideCollapsedHandle() {
+    if (_isCollapsedHandleAutoHidden) {
+      _isCollapsedHandleAutoHidden = false;
+      notifyListeners();
+    }
+  }
+}
