+++ super_editor/lib/super_editor.dart
@@ -17,7 +17,7 @@ export 'src/default_editor/blockquote.dart';
 export 'src/default_editor/box_component.dart';
 export 'src/default_editor/common_editor_operations.dart';
 export 'src/default_editor/document_caret_overlay.dart';
-export 'src/default_editor/document_gestures.dart';
+export 'src/infrastructure/document_gestures.dart';
 export 'src/default_editor/document_gestures_mouse.dart';
 export 'src/default_editor/document_gestures_touch.dart';
 export 'src/default_editor/document_input_ime.dart';
@@ -42,6 +42,9 @@ export 'src/infrastructure/composable_text.dart';
 export 'src/infrastructure/focus.dart';
 export 'src/infrastructure/ime_input_owner.dart';
 export 'src/infrastructure/multi_tap_gesture.dart';
+export 'src/infrastructure/platforms/android/android_document_controls.dart';
+export 'src/infrastructure/platforms/ios/ios_document_controls.dart';
+export 'src/infrastructure/platforms/mobile_documents.dart';
 export 'src/infrastructure/scrolling_diagnostics/scrolling_diagnostics.dart';
 export 'src/infrastructure/strings.dart';
 export 'src/infrastructure/super_textfield/super_textfield.dart';
