+++ super_editor/test/super_textfield/toolbar_position_delegate_test.dart
@@ -1,7 +1,7 @@
 import 'dart:ui';
 
 import 'package:flutter_test/flutter_test.dart';
-import 'package:super_editor/src/infrastructure/super_textfield/infrastructure/toolbar_position_delegate.dart';
+import 'package:super_editor/src/infrastructure/toolbar_position_delegate.dart';
 
 void main() {
   group('SuperTextField', () {
